Extract parent dataset utils
