+++ src/jlmkr/utils/parent_dataset.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: © 2024 Jip-Hop and the Jailmakers <https://github.com/Jip-Hop/jailmaker>
+#
+# SPDX-License-Identifier: LGPL-3.0-only
+
+import os
+
+from donor.jlmkr import JAILS_DIR_PATH
+
+
+def get_all_jail_names():
+    try:
+        jail_names = os.listdir(JAILS_DIR_PATH)
+    except FileNotFoundError:
+        jail_names = []
+
+    return jail_names
